refactoring code for the 4D model to avoid duplicate code
