+++ src/ifc4d/ifc4d/msp2ifc.py
@@ -23,6 +23,7 @@ import ifcopenshell
 import ifcopenshell.api
 import ifcopenshell.util.date
 import xml.etree.ElementTree as ET
+from common import Utils
 
 
 class MSP2Ifc:
@@ -40,7 +41,10 @@ class MSP2Ifc:
 
     def execute(self):
         self.parse_xml()
-        self.create_ifc()
+        ifcCreator = Utils(self.file, self.work_plan, self.project, self.calendars,
+                           self.wbs, self.root_activites, self.activities, self.relationships)
+        ifcCreator.create_ifc()
+        #self.create_ifc()
 
     def parse_xml(self):
         tree = ET.parse(self.xml)
@@ -120,135 +124,3 @@ class MSP2Ifc:
                 "Name": calendar.find("pr:Name", self.ns).text,
                 "StandardWorkWeek": week_days,
             }
-
-    def create_ifc(self):
-        if not self.file:
-            self.create_boilerplate_ifc()
-        if not self.work_plan:
-            self.work_plan = ifcopenshell.api.run("sequence.add_work_plan", self.file)
-        work_schedule = self.create_work_schedule()
-        self.create_tasks(work_schedule)
-        self.create_calendars()
-        self.create_rel_sequences()
-
-    def create_boilerplate_ifc(self):
-        self.file = ifcopenshell.file(schema="IFC4")
-        self.work_plan = self.file.create_entity("IfcWorkPlan")
-
-    def create_tasks(self, work_schedule):
-        for task_id in self.tasks:
-            task = self.tasks[task_id]
-            if task["OutlineLevel"] == 0:
-                self.create_task(task, work_schedule=work_schedule)
-
-    def create_work_schedule(self):
-        return ifcopenshell.api.run(
-            "sequence.add_work_schedule", self.file, name=self.project["Name"], work_plan=self.work_plan
-        )
-
-    def create_calendars(self):
-        for calendar in self.calendars.values():
-            calendar["ifc"] = ifcopenshell.api.run("sequence.add_work_calendar", self.file, name=calendar["Name"])
-            self.process_working_week(calendar["StandardWorkWeek"], calendar["ifc"])
-
-    def create_task(self, task, work_schedule=None, parent_task=None):
-        task["ifc"] = ifcopenshell.api.run(
-            "sequence.add_task",
-            self.file,
-            work_schedule=work_schedule if work_schedule else None,
-            parent_task=parent_task["ifc"] if parent_task else None,
-        )
-        ifcopenshell.api.run(
-            "sequence.edit_task",
-            self.file,
-            task=task["ifc"],
-            attributes={
-                "Name": task["Name"],
-                "Identification": task["OutlineNumber"],
-                "IsMilestone": task["Start"] == task["Finish"],
-            },
-        )
-        task_time = ifcopenshell.api.run("sequence.add_task_time", self.file, task=task["ifc"])
-        ifcopenshell.api.run(
-            "sequence.edit_task_time",
-            self.file,
-            task_time=task_time,
-            attributes={
-                "ScheduleStart": task["Start"],
-                "ScheduleFinish": task["Finish"],
-                "DurationType": "WORKTIME" if task["Duration"] else None,
-                "ScheduleDuration": task["Duration"] if task["Duration"] else None,
-            },
-        )
-        for subtask_id in task["subtasks"]:
-            self.create_task(self.tasks[subtask_id], parent_task=task)
-
-    def process_working_week(self, week, calendar):
-        for day in week:
-            if day["ifc"]:
-                continue
-
-            day["ifc"] = ifcopenshell.api.run(
-                "sequence.add_work_time", self.file, work_calendar=calendar, time_type="WorkingTimes"
-            )
-
-            weekday_component = [int(day["DayType"])]
-            for day2 in week:
-                if day["DayType"] == day2["DayType"]:
-                    continue
-                if day["WorkingTimes"] == day2["WorkingTimes"]:
-                    weekday_component.append(int(day2["DayType"]))
-                    # Don't process the next day, as we can group it
-                    day2["ifc"] = day["ifc"]
-
-            work_time_name = "Weekdays: {}".format(", ".join([str(c) for c in sorted(weekday_component)]))
-            ifcopenshell.api.run(
-                "sequence.edit_work_time",
-                self.file,
-                work_time=day["ifc"],
-                attributes={"Name": work_time_name},
-            )
-
-            recurrence = ifcopenshell.api.run(
-                "sequence.assign_recurrence_pattern", self.file, parent=day["ifc"], recurrence_type="WEEKLY"
-            )
-            ifcopenshell.api.run(
-                "sequence.edit_recurrence_pattern",
-                self.file,
-                recurrence_pattern=recurrence,
-                attributes={"WeekdayComponent": weekday_component},
-            )
-            for work_time in day["WorkingTimes"]:
-                ifcopenshell.api.run(
-                    "sequence.add_time_period",
-                    self.file,
-                    recurrence_pattern=recurrence,
-                    start_time=work_time["Start"],
-                    end_time=work_time["Finish"],
-                )
-
-    def create_rel_sequences(self):
-        self.sequence_type_map = {
-            "1": "START_START",
-            "2": "START_FINISH",
-            "3": "FINISH_START",
-            "4": "FINISH_FINISH",
-            "0": "NOTDEFINED",
-        }
-        for task in self.tasks.values():
-            if not task["PredecessorTasks"]:
-                continue
-            for predecessor in task["PredecessorTasks"].values():
-                rel_sequence = ifcopenshell.api.run(
-                    "sequence.assign_sequence",
-                    self.file,
-                    related_process=task["ifc"],
-                    relating_process=self.tasks[predecessor["PredecessorTask"]]["ifc"],
-                )
-                if predecessor["Type"]:
-                    ifcopenshell.api.run(
-                        "sequence.edit_sequence",
-                        self.file,
-                        rel_sequence=rel_sequence,
-                        attributes={"SequenceType": self.sequence_type_map[predecessor["Type"]]},
-                    )
